// libs/package-management/src/project/project-types.ts
export interface PackageDependencyItem {
  name: string;
  type: PackageDependencyType;
  version: string;
}

export type PackageDependencyType =
  | "dependency"
  | "devDependency"
  | "peerDependency"
  | "optionalDependency";

export const dependencyTypeMap = {
  dependency: "dependencies",
  devDependency: "devDependencies",
  peerDependency: "peerDependencies",
  optionalDependency: "optionalDependencies",
} satisfies Record<PackageDependencyType, keyof KnownPackageJson>;

export type KnownPackageJson = OmitIndexSignature<PackageJson>;